# doc/a3s-openapi.yaml
openapi: 3.0.0
info:
  description: >-
    API Definition for A3S. This service allows authentication, authorisation and accounting.
  version: 1.1.3
  title: A3S
paths:
...
  '/permissions':
    get:
      tags:
        - Permission
      description: Search for permissions.
      summary: Search for permissions.
      operationId: listPermissions
      parameters:
        - name: page
          in: query
          required: false
          description: The page to view.
          schema:
            type: integer
            minimum: 1
            format: int32
            default: 1
        - name: size
          in: query
          required: false
          description: The size of a page.
          schema:
            type: integer
            minimum: 1
            maximum: 20
            format: int32
            default: 10
        - name: filterName
          in: query
          required: false
          description: A search query filter on the permission's name.
          schema:
            type: string
            minLength: 0
            maxLength: 255
          example: FilterByEntityType
        - name: orderBy
          in: query
          required: false
          style: form
          explode: false
          description: >-
            a comma separated list of fields in their sort order. Ascending
            order is assumed. Append '_desc' after a field to indicate descending order. Supported fields. 'name'.
          schema:
            type: string
      responses:
        '200':
          description: OK
          headers:
            X-Pagination:
              description: Pagination header.
              schema:
                $ref: '#/components/schemas/PaginationHeaderResponse'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Permission'
        '400':
          description: Bad Request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - You are not authorized to access the list of permissions.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Permission list not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        500:
          description: An unexpected error occurred.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  '/permissions/{permissionId}':
    get:
      tags:
        - Permission
      summary: Get a permission.
      description: Get a permission by its UUID.
      operationId: getPermission
      parameters:
        - name: permissionId
          in: path
          description: permission
          required: true
          schema:
            $ref: '#/components/schemas/PermissionId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Permission'
        '400':
          description: Bad Request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - You are not authorized to access the permission.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Permission not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        500:
          description: An unexpected error occurred
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  #endregion
  
  #region - TermsOfService
...
components:
  schemas:
...
    PermissionId:
      type: string
      format: uuid
    Permission:
      description: |
        Permission for an application
      type: object
      properties:
        uuid:
          $ref: '#/components/schemas/PermissionId'
        name:
          type: string
        description:
          type: string

    # End Region - Permissions      

    # Region - TermsOfService
...
    ErrorResponse:
      description: Error response object to give more info with HTTP status code
      type: object
      required:
        - message
      properties:
        message:
          type: string
    PaginationHeaderResponse:
      description: >-
        This is the standard response for requests that have paginated
        collections
      type: object
      properties:
        total:
          type: integer
          format: int32
          minimum: 1
          description: The total number of pages in the result set
        count:
          type: integer
          format: int32
          minimum: 1
          description: The total number of results in the result set
        size:
          type: integer
          format: int32
          minimum: 1
          maximum: 20
          default: 10
          description: >-
            The number items to include in a page of results. The page size for
            the result set
        current:
          type: integer
          format: int32
          minimum: 1
          description: >-
            The position of the page in the paged result set that is being
            returned
        first:
          type: string
          format: url
          description: >-
            The link to the first page of results containing [size] results.
            This link includes the page, size, filter, orderBy and fields query
            parameters.
        last:
          type: string
          format: url
          description: >-
            The link to the last page of results containing [size] results. This
            link includes the page, size, filter, orderBy and fields query
            parameters.
        prev:
          type: string
          format: url
          description: >-
            The link to the previous page of results containing [size] results.
            This link includes the page, size, filter, orderBy and fields query
            parameters. If this is the first page in the result set then this
            will be [null].
        next:
          type: string
          format: url
          description: >-
            The link to the next page of results containing [size] results. This
            link includes the page, size, filter, orderBy and fields query
            parameters. If this is the last page in the result set then this will
            be [null].
...